Created bar_search method
The bar search method takes an address and distance (miles), and returns a list of bars that are within the distance of the given address. Inside the bar list you will also find each bar's rating and address. The information comes from the Google Places API. I've also created a geocoder to get the latitude and longitude coords of the address (info also comes from Google Places API). There is also a method to parse through the returned geocoder data, and a simple miles to meters converter (written mainly to check that the given distance is within the correct bounds).